refactor(apollo-react): rename collapse.ts to node-size.ts
The file only holds the shape to default-size helper now (getExpandedSize),
not collapse logic. Renamed for clarity; export names kept so consumers
(flow-workbench) don't break.

BREAKING CHANGE: getCollapsedShape, getCollapsedSize, and getExpandedShape are
removed from the public canvas/utils API. They only supported the removed
collapse-to-shape tying (collapse no longer changes a node's shape).
getExpandedSize and the size constants are unaffected.
