+++ packages/apollo-react/src/canvas/components/AddNodePanel/AddNodeManager.helpers.test.ts
@@ -3,7 +3,7 @@ import { describe, expect, it } from 'vitest';
 import { PREVIEW_NODE_ID } from '../../constants';
 import type { NodeLayout } from '../../hooks/useCanvasNodeLayout';
 import type { NodeShape } from '../../schema/node-definition';
-import { getExpandedSize } from '../../utils/collapse';
+import { getExpandedSize } from '../../utils/node-size';
 import { getContainerFitGeometry, getNodeDimensions } from '../../utils/container';
 import { placeAddedNode } from './AddNodeManager.helpers';
