+++ packages/apollo-react/src/canvas/hooks/useCanvasNodeLayout.ts
@@ -2,7 +2,7 @@ import type { Node } from '@uipath/apollo-react/canvas/xyflow/react';
 import { useCallback, useMemo } from 'react';
 import { useOptionalNodeTypeRegistry } from '../core';
 import type { NodeManifest } from '../schema/node-definition';
-import { getExpandedSize } from '../utils/collapse';
+import { getExpandedSize } from '../utils/node-size';
 import {
   type ContainerFitGeometry,
   getContainerFitGeometry as getDefaultContainerFitGeometry,
